+++ frontend/aimposter/src/components/roles/MafiaScreen.js
@@ -1,10 +1,10 @@
-// src/components/HumanRoleScreen.js
+// src/components/MafiaScreen.js
 import React from 'react';
 import './MafiaScreen.css';
-import human from '../assets/images/human.png';
-import logo from '../assets/images/logo.png';
+import robot from '../../assets/images/mafia_ai.png';
+import logo from '../../assets/images/logo.png';
 
-const HumanRoleScreen = () => {
+const MafiaScreen = () => {
   return (
     <div className="welcome-screen">
       <div className="header">
@@ -17,13 +17,13 @@ const HumanRoleScreen = () => {
       </div>
       <div className="text">
         <h1>ПОЗДРАВЛЯЕМ! ВАША РОЛЬ</h1>
-        <h2 className="mafia-ai">ЧЕЛОВЕК</h2>
+        <h2 className="mafia-ai">МАФИЯ AI</h2>
       </div>
       <div className="robot-icon">
-        <img src={human} alt="Human" />
+        <img src={robot} alt="Robot" />
       </div>
     </div>
   );
 };
 
-export default HumanRoleScreen;
+export default MafiaScreen;
